+++ src/bridge.js
@@ -87,9 +87,12 @@ export default class Bridge {
     }
 
     const beforeStreamRenderCallback = () => {
+      // In theory, 2 frames would be enough for everyone but Safari
       requestAnimationFrame(() => {
         requestAnimationFrame(() => {
-          initCallback()
+          requestAnimationFrame(() => {
+            initCallback()
+          })
         })
       })
     }